create data base

+++ app/src/main/java/com/example/myapplication/entity/CountryCard.kt
@@ -2,6 +2,6 @@ package com.example.myapplication.entity
 
 interface CountryCard {
     val name: String?
-    val latitude: Long?
-    val longitude: Long?
+    val latitude: Double?
+    val longitude: Double?
 }
